refactor: state backends and PR review follow-ups (ENG-9730)
- Split HTTP, local file, and S3 into state_backends with load/save
- Add resolve_state_backend_config; tidy duplicate S3/path resolution
- Move pyproject discovery to project_discovery to avoid import cycles
- Pydantic StateBackendConfig; re-export legacy names via state_storage
- CLI: clearer state backend debug messages
- S3: top-level boto3/botocore imports with lazy backend import
- Tests: moto for S3; patch integration filter at new module path
