+++ pyproject.toml
@@ -12,6 +12,7 @@ dev = [
     "basedpyright",
     "boto3",
     "dbt-core>=1.10,<1.12",
+    "moto[s3]",
     "pytest-httpx",
     "pytest",
     "ruff",
