+++ src/orchestra_dbt/state_backends/base.py
@@ -1,0 +1,9 @@
+from typing import Protocol
+
+from ..models import StateApiModel
+
+
+class StateBackend(Protocol):
+    def load(self) -> StateApiModel: ...
+
+    def save(self, state: StateApiModel) -> None: ...
